requirements.txt
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.1; python_full_version >= '3.7.0'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.40; python_version >= '3.7'
giturlparse==0.12.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
python-gitlab==3.15.0; python_full_version >= '3.7.0'
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
urllib3==2.0.7; python_version >= '3.7'
